requirements.txt
attrs==19.3.0
click==7.1.2
Flask==1.1.2
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
